+++ core/build.gradle.kts
@@ -1,11 +1,9 @@
 plugins {
     `kotlin-script`
     `adventure-script`
-    id("fabric-loom")
+    `fabric-script`
 }
 
 dependencies {
-    val gameVersion: String by properties
-    minecraft("com.mojang", "minecraft", gameVersion)
-    mappings(loom.officialMojangMappings())
+
 }
