+++ core/src/main/kotlin/de/miraculixx/veinminer/config/UpdateManager.kt
@@ -2,6 +2,7 @@ package de.miraculixx.veinminer.config
 
 import de.miraculixx.veinminer.config.data.ModrinthFile
 import de.miraculixx.veinminer.config.data.ModrinthVersion
+import de.miraculixx.veinminer.config.utils.json
 import java.io.File
 import java.net.HttpURLConnection
 import java.net.URI
