+++ core/src/main/kotlin/de/miraculixx/veinminer/config/extensions/PathExtensions.kt
@@ -1,5 +1,6 @@
-package de.miraculixx.veinminer.config
+package de.miraculixx.veinminer.config.extensions
 
+import de.miraculixx.veinminer.config.utils.json
 import net.kyori.adventure.text.format.TextColor
 import java.nio.file.Path
 import kotlin.io.path.createFile
