+++ veinminer-client/build.gradle.kts
@@ -6,7 +6,7 @@ plugins {
 }
 
 dependencies {
-    implementation(include(project(":core"))!!)
+    implementation(include(project(":core", configuration = "namedElements"))!!)
 }
