+++ veinminer-client/src/main/kotlin/de/miraculixx/veinminerClient/network/NetworkManager.kt
@@ -3,11 +3,14 @@ package de.miraculixx.veinminerClient.network
 import de.miraculixx.veinminer.config.data.FixedBlockGroup
 import de.miraculixx.veinminer.config.extensions.toVeinminer
 import de.miraculixx.veinminer.config.network.JoinInformation
+import de.miraculixx.veinminer.config.network.KeyPress
 import de.miraculixx.veinminer.config.network.RequestBlockVein
 import de.miraculixx.veinminer.config.pattern.Pattern
+import de.miraculixx.veinminerClient.VeinminerClient
 import de.miraculixx.veinminerClient.constants.PACKET_CONFIGURATION
 import de.miraculixx.veinminerClient.constants.PACKET_HIGHLIGHT
 import de.miraculixx.veinminerClient.constants.PACKET_JOIN
+import de.miraculixx.veinminerClient.constants.PACKET_KEY_PRESS
 import de.miraculixx.veinminerClient.constants.PACKET_MINE
 import de.miraculixx.veinminerClient.render.BlockHighlightingRenderer
 import net.minecraft.client.gui.components.toasts.SystemToast
@@ -46,6 +49,7 @@ object NetworkManager {
     }
 
     private val onHighlight = PACKET_HIGHLIGHT.receiveOnClient { packet, context ->
+        VeinminerClient.LOGGER.info("Received block highlight: $packet")
         if (!packet.allowed) {
             // TODO
             return@receiveOnClient
@@ -56,14 +60,17 @@ object NetworkManager {
     }
 
     fun requestBlockInfo(position: BlockPos, direction: Direction) {
-        PACKET_MINE.send(RequestBlockVein(position.toVeinminer(), direction.toVeinminer(), selectedPattern, false))
-    }
-
-    fun requestBlockMine(position: BlockPos, direction: Direction) {
-        PACKET_MINE.send(RequestBlockVein(position.toVeinminer(), direction.toVeinminer(), selectedPattern, true))
+        VeinminerClient.LOGGER.info("Sending veinmine request: ($position, $direction)")
+        PACKET_MINE.send(RequestBlockVein(position.toVeinminer(), direction.toVeinminer(), selectedPattern))
     }
 
     fun sendJoin(version: String) {
+        VeinminerClient.LOGGER.info("Sending join: ($version)")
         PACKET_JOIN.send(JoinInformation(version))
+    }
+
+    fun sendKeyPress(pressed: Boolean) {
+        VeinminerClient.LOGGER.info("Sending key press: $pressed")
+        PACKET_KEY_PRESS.send(KeyPress(pressed))
     }
 }
